DRY up video service auth code
